% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect2ls.R
\name{connect2ls}
\alias{connect2ls}
\title{Connect to LimeSurvey API}
\usage{
connect2ls(
  lime_username,
  lime_password,
  url = "https://badania.stocznia.org.pl/index.php?r=admin/remotecontrol"
)
}
\arguments{
\item{lime_username}{Your LimeSurvey username.}

\item{lime_password}{Your LimeSurvey password.}

\item{url}{LimeSurvey RemoteControl API 2 url.}
}
\description{
\code{connect2ls()} allows you to connect to LimeSurvey RemoteControl API.
It uses a modified version of the \code{limer} library to establish connection.
}
\details{
The creators of \code{limer} no longer maintain it.
An up-to-date version of the package, compatible with the latest versions of LimeSurvey,
can be installed directly from my github \code{devtools::install_github("wjk-g/limer")}.
}
