% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_excel.R
\name{export_to_excel}
\alias{export_to_excel}
\title{Export data to Excel}
\usage{
export_to_excel(
  data,
  varnames = list(),
  max_cat = 15,
  bonferroni = FALSE,
  filename = paste0(deparse(substitute(data)), "_tables.xlsx"),
  overwrite = FALSE
)
}
\arguments{
\item{data}{A dataframe}

\item{varnames}{Variables}

\item{max_cat}{The maximum number of categories for a variable}

\item{bonferroni}{Bonferroni correction}

\item{filename}{Filename}

\item{overwrite}{Overwrite}
}
\description{
Export data to Excel
}
