% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_labels_from_data.R
\name{extract_labels_from_data}
\alias{extract_labels_from_data}
\title{Extract question labels from LimeSurvey}
\usage{
extract_labels_from_data(labelled_data)
}
\arguments{
\item{labelled_data}{A path to a .sav file or a labelled R dataframe.}
}
\description{
The function extracts variable descriptions directly from an SPSS file
(if a path is provided) or from a dataframe object with labels
attached (i.e. with a \verb{$variable.labels} attribute).
}
