% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_labels_from_ls.R
\name{extract_labels_from_ls}
\alias{extract_labels_from_ls}
\title{Extract question labels from LimeSurvey}
\usage{
extract_labels_from_ls(svyid)
}
\arguments{
\item{svyid}{id number of a LimeSurvey survey.}
}
\description{
The function extracts variable descriptions from LimeSurvey.
}
\details{
Connecting to LimeSurvey and downloading the necessary data can
take between 10 and 20 seconds, depending on a survey.
}
\section{Example}{
\preformatted{connect2ls('your_username', 'your_password', 'ls_url')

# Import without assigning to variable
extract_labels_from_ls(survey_number)

# Import and assign to a variable (useful when working with multiple surveys)
labels_for_survey_1 <- extract_labels_from_ls(survey_1_number)
labels_for_survey_2 <- extract_labels_from_ls(survey_2_number)

}
}

