% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colnames.R
\name{get_colnames}
\alias{get_colnames}
\title{Get variable descriptions}
\usage{
get_colnames(df, source = enlabel_cache$codes_and_labels, trim = FALSE)
}
\arguments{
\item{df}{A dataframe to be labelled.}

\item{source}{A dataframe containing question labels.}

\item{trim}{If \code{TRUE}, \code{get_colnames()} will attempt to include only the subquestion text
and discard the parent-question text.}
}
\description{
This functions takes a dataframe as an input and returns a vector of variable descriptions.
You can use \code{get_colnames()} boh on 'unspoiled' and modified dataframes,
i.e. LimeSurvey or SPSS survey files that have had variables removed, renamed,
added or otherwise changed. If a variable does not have a match in the dataframe
supplied to \code{source}, its name will be used as a label.

The main goal of \code{get_colnames()} is to prepare your data for export, e.g.
to Excel with the \code{expss} package.
}
\details{
Produces a character vector of human-readable question labels
}
