% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_variables.R
\name{label_variables}
\alias{label_variables}
\title{Attach labels to variable names}
\usage{
label_variables(df, labels)
}
\arguments{
\item{df}{A df to be labelled.}

\item{labels}{A character vector containing variable labels.}
}
\description{
Attach labels to variable names
}
